requirements.txt
Flask==2.1.3
Flask-API==3.0.post1
PyJWT==1.7.1
requests==2.28.1
psutil==5.9.4


